started material assignment stuff

+++ blender_portion.py
@@ -2,20 +2,26 @@ import sys
 import argparse
 import logging
 import time
+import json
+import imp
+
 import bpy
+import bpy.py.utils as utils
+
+imp.reload(utils)
 
 thisLogger = logging.getLogger("BlenderScriptLogger")
 thisLogger.setLevel(logging.DEBUG)
 
 HIGHLIGHT_NAMES = ["rig:margo_bun_geo", "rig:margo_hair_geo"]
-NOHIGHLIGHT_NAMES = ["rig:noseLeft_geo", "rig:mouthHighResBase",
-"rig:margo_body_geo", "rig:eyeRight_geo_mask", "rig:eyeRight"]
+NOHIGHLIGHT_NAMES = ["rig:mouthHighResBase","rig:margo_body_geo"]
 EYE_NAMES = ['rig:eyeLeft_geo', 'rig:eyeRight_geo']
 EYEMASK_NAMES = ["rig:eyeLeft_geo_mask", "rig:eyeRight_geo_mask"]
 FACE_NAMES = ["rig:mouthHighResBase", "rig:noseLeft_geo"]
 
-collections = {HIGHLIGHT_NAMES:'highlight_geo', NOHIGHLIGHT_NAMES:'nohighlight',
-EYEMASK_NAMES:'eye_mask', EYE_NAMES:'eyes', FACE_NAMES:'face'}
+rigObjs = {0:HIGHLIGHT_NAMES, 1:NOHIGHLIGHT_NAMES, 2:EYEMASK_NAMES, 3:EYE_NAMES,
+4:FACE_NAMES}
+sceneCollections = {0:'highlight_geo', 1:'nohighlight', 2:'eye_mask', 3:'eyes', 4:'face'}
 
 def get_args():
     if '--' in sys.argv:
@@ -36,6 +42,7 @@ def import_abc(abcPath):
 
 def sort_objects():
     objs = bpy.data.objects
+    thisLogger.info("Objects are %s", objs)
     cols = bpy.data.collections
     pd = {}
     #Make a dictionary of objects:highest level parent
@@ -54,30 +61,52 @@ def sort_objects():
                 cc = c
                 break
         if "2d" in p2.name.lower():
-            cols['2d'].objects.link(obj)
-            cc.objects.unlink(obj)
+            if cc != cols['2d']:
+                cols['2d'].objects.link(obj)
+                cc.objects.unlink(obj)
+            continue
         if "3d" in p2.name.lower():
-            cols['nohighlight'].objects.link(obj)
-            cc.objects.unlink(obj)
-        for k in collections.keys():
-            if p2.name in k:
-                col = collections[k]
-                cols[col].objects.link(obj)
-                cc.object.unlink(obj)
+            if cc != cols['nohighlight']:
+                cols['nohighlight'].objects.link(obj)
+                cc.objects.unlink(obj)
+            continue
+        if "camera" in p2.name.lower():
+            if cc != cols['cameras']:
+                cols['cameras'].objects.link(obj)
+                cc.objects.unlink(obj)
+            continue
+        for k in rigObjs.keys():
+            if p2.name in rigObjs[k]:
+                col = sceneCollections[k]
+                if cc != col:
+                    cols[col].objects.link(obj)
+                    cc.objects.unlink(obj)
                 break
             else:
                 continue
 
+def apply_materials(path):
+    f = open(path)
+    matData = json.load(f)
+    for obj in matData:
+        for mat in matData[obj]:
+            m = None
+            try:
+                m = bpy.data.materials[mat]
+            except:
+                thisLogger.info("Creating Material for %s" % mat)
+                m = utils.create_material_from_dict(matData[obj][mat])
+            utils.apply_material_to_object(obj, m)
+            
 
 
-
-
-
+    
 
 p1, p2 = get_args()
 print("path1 is ", p1)
 print("path2 is ", p2)
 import_abc(p2)
+sort_objects()
 
 """
 Collections:
